// src/app/api/model/restaurante.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Plato } from './plato';
import { Direccion } from './direccion';


export interface Restaurante { 
    id?: string;
    nombre?: string;
    description?: string;
    iconUrls?: Array<string>;
    tags?: Array<string>;
    platos?: Set<Plato>;
    direcciones?: Array<Direccion>;
    rating?: number;
}